Feature/error handler (#7)
* create and add global error handler

* add not found error handler

* add bad request error handler

* create and add ErrorResponse

* resource not found error implemented

---------

+++ api-task/src/main/java/com/noeliacarosella/apitask/service/impl/TaskService.java
@@ -1,5 +1,6 @@
 package com.noeliacarosella.apitask.service.impl;
 
+import com.noeliacarosella.apitask.exceptions.ResourceNotFoundException;
 import com.noeliacarosella.apitask.model.Task;
 import com.noeliacarosella.apitask.repository.TaskRepository;
 import com.noeliacarosella.apitask.service.ITaskService;
@@ -7,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TaskService implements ITaskService {
@@ -25,12 +27,22 @@ public class TaskService implements ITaskService {
     }
 
     @Override
-    public Task update(Task task) {
-        return taskRepository.save(task);
+    public Task update(Task task) throws ResourceNotFoundException {
+        Optional<Task> existingTask = taskRepository.findById(task.getId());
+        if (existingTask.isPresent()) {
+            return taskRepository.save(task);
+        } else {
+            throw new ResourceNotFoundException("Task Not Found");
+        }
     }
 
     @Override
-    public void delete(Long id) {
-        taskRepository.deleteById(id);
+    public void delete(Long id) throws ResourceNotFoundException {
+        Optional<Task> existingTask = taskRepository.findById(id);
+        if (existingTask.isPresent()) {
+            taskRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("Task Not Found");
+        }
     }
 }
